pages/main_page.py
from .base_page import BasePage
from selenium.webdriver.common.by import By
from .locators import MainPageLocators
from .locators import BookPageLocators
from .login_page import LoginPage


class MainPage(BasePage):
    def go_to_login_page(self):
        login_link = self.browser.find_element(*MainPageLocators.LOGIN_LINK)
        login_link.click()

    def should_be_login_link(self):
        assert self.is_element_present(*MainPageLocators.LOGIN_LINK), "Login link is not presented"

    def add_to_cart(self):
        try:
            self.browser.find_element(*BookPageLocators.CART_BUTTON).click()
        except:
            assert False, 'Кнопка корзины не найдена'
        assert True






